# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CaseStudy API",
        "description": "Finance management API with JWT authentication",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/CaseStudy/api/public/api",
            "description": "Local XAMPP server"
        }
    ],
    "paths": {
...
        "/budgets": {
            "get": {
                "tags": [
                    "Budgets"
                ],
                "summary": "List all budgets for the authenticated user",
                "operationId": "49a5c108bcc6c0f36d6721daf5ff8732",
                "responses": {
                    "200": {
                        "description": "List of budgets with category names"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Budgets"
                ],
                "summary": "Create a new budget",
                "operationId": "99b08fee5d3d4d75f33713e9343f0c91",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "category_id",
                                    "budget_month",
                                    "budget_year",
                                    "budget_amount"
                                ],
                                "properties": {
                                    "category_id": {
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "budget_month": {
                                        "type": "integer",
                                        "example": 6,
                                        "maximum": 12,
                                        "minimum": 1
                                    },
                                    "budget_year": {
                                        "type": "integer",
                                        "example": 2024,
                                        "minimum": 2000
                                    },
                                    "budget_amount": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 5000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Budget Saved Successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/budgets/{id}": {
            "put": {
                "tags": [
                    "Budgets"
                ],
                "summary": "Update a budget",
                "operationId": "dfac0f7e44456370636bfda89a9403ae",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "category_id",
                                    "budget_month",
                                    "budget_year",
                                    "budget_amount"
                                ],
                                "properties": {
                                    "category_id": {
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "budget_month": {
                                        "type": "integer",
                                        "example": 6,
                                        "maximum": 12,
                                        "minimum": 1
                                    },
                                    "budget_year": {
                                        "type": "integer",
                                        "example": 2024,
                                        "minimum": 2000
                                    },
                                    "budget_amount": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 6000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Budget Updated Successfully"
                    },
                    "404": {
                        "description": "Budget Not Found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Budgets"
                ],
                "summary": "Delete a budget",
                "operationId": "46c1727c8456d5ae253255b278b71ffd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Budget Deleted Successfully"
                    },
                    "404": {
                        "description": "Budget Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/budget-tracking": {
            "get": {
                "tags": [
                    "Budgets"
                ],
                "summary": "Get budget tracking with spending vs budget per category",
                "operationId": "13ebb25e87ba5ed7cc0e4ff8e6e20c56",
                "responses": {
                    "200": {
                        "description": "Budget tracking data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "category_name": {
                                                "type": "string"
                                            },
                                            "budget_amount": {
                                                "type": "number"
                                            },
                                            "spent_amount": {
                                                "type": "number"
                                            },
                                            "percentage": {
                                                "type": "number"
                                            },
                                            "status": {
                                                "type": "string",
                                                "enum": [
                                                    "Safe",
                                                    "Warning",
                                                    "Exceeded"
                                                ]
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/budget-overview": {
            "get": {
                "tags": [
                    "Budgets"
                ],
                "summary": "Get overall budget overview for the authenticated user",
                "operationId": "4402a1b34c344ef3872d578516606b2a",
                "responses": {
                    "200": {
                        "description": "Budget overview",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "budget": {
                                            "type": "number"
                                        },
                                        "spent": {
                                            "type": "number"
                                        },
                                        "remaining": {
                                            "type": "number"
                                        },
                                        "status": {
                                            "type": "string",
                                            "enum": [
                                                "Safe",
                                                "Warning",
                                                "Exceeded"
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}